Enable dotenv for test and move environment config
Gems and Rails initializers are loaded by the rspec script before
rails_helper.rb, so environment variables set in this file cannot be
used in initializers.

As a workaround and to achieve the same effect, this enables dotenv for
the test environment as well, and moves the environment variables from
rails_helper.rb to .env.test.